rename commit_preprocessor package to commit_processor
