+++ prospector/commit_processor/Idea_1.png
