+++ prospector/commit_processor/SLAB Ideas.md
