+++ prospector/commit_processor/constants.py
